requirements.txt
google-adk>=1.27.0
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
